Adds check for dir before deleting data

+++ prime-router/build.gradle.kts
@@ -111,8 +111,10 @@ tasks.clean {
     // clean up all the old event files in the SOAP set up
     doLast {
         val eventsDir = File("../.environment/soap_service/soap/event/v1/")
-        FileUtils.listFiles(eventsDir, arrayOf("event"), true).forEach {
-            it.delete()
+        if (eventsDir.exists()) {
+            FileUtils.listFiles(eventsDir, arrayOf("event"), true).forEach {
+                it.delete()
+            }
         }
     }
 }
